workaround for GSH hanging bug.
when a signal is received, it's temporarily blocked while the signal function is called.

if SIGCHILD is blocked while waiting in wait(), the SIGCHILD is pended and the wait() wakeup machinery never happens and the wait() call never returns.

I'm not entirely sure how it gets into that state to begin with, but this flags the wait() before the signal is blocked.